requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.36
asyncpg==0.29.0
alembic==1.13.3

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# Task queue
rq==1.16.2
redis==5.0.8

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
aiosqlite==0.20.0

# Utils
pydantic[email]==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1